requirements.txt
appdirs==1.4.4
attrs==21.2.0
boto3==1.18.44
botocore==1.21.44
CacheControl==0.12.6
cachetools==4.2.2
cattrs==1.8.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
DateTime==4.3
google-api-core==2.0.1
google-auth==2.1.0
google-cloud==0.34.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
idna==3.2
iso3166==2.0.1
jmespath==0.10.0
mapbox==0.18.0
msgpack==1.0.2
numpy==1.21.2
pandas==1.3.3
polyline==1.4.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-cache==0.8.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
uritemplate==3.0.1
url-normalize==1.4.3
urllib3==1.26.6
zope.interface==5.4.0
